local-ui-analyzer/requirements.txt
ollama
opencv-python
numpy
scipy
jinja2
matplotlib
requests
playwright
google-genai
python-dotenv

torch
torchvision
albumentations
tqdm
Pillow

